+++ apsg/__init__.py
@@ -3,7 +3,8 @@
 from .core import Vec3, Fol, Lin, Pair, Fault
 from .core import Group, FaultSet, Ortensor, Cluster, G
 from .plotting import StereoNet
+from .db import SDB
 
 __author__ = 'Ondrej Lexa'
 __email__ = 'lexa.ondrej@gmail.com'
-__version__ = '0.3.0'
+__version__ = '0.3.1'