+++ apsg/db.py
@@ -7,6 +7,8 @@ API to read data from PySDB database
 import sqlite3
 from .core import Fol, Lin, Group
 
+__all__ = ['SDB']
+
 
 class SDB(object):
     """PySDB database access class"""